updated writers, logging

+++ training.py
@@ -18,18 +18,16 @@ import random
 import yaml
 import logging
 import shutil
-
+from models.resnet import *
 from utils.utils import *
 from utils.image_helper import ImageHelper
 from utils.text_helper import TextHelper
-logger = logging.getLogger('logger')
 from prompt_toolkit import prompt
 
+logger = logging.getLogger('logger')
 
 
-
-
-def train(run_helper: ImageHelper, model: nn.Module, optimizer, criterion, writer, epoch):
+def train(run_helper: ImageHelper, model: nn.Module, optimizer, criterion, epoch):
     train_loader = run_helper.train_loader
     model.train()
     running_loss = 0.0
@@ -40,6 +38,10 @@ def train(run_helper: ImageHelper, model: nn.Module, optimizer, criterion, write
         inputs = inputs.to(run_helper.device)
         labels = labels.to(run_helper.device)
         # zero the parameter gradients
+
+        if helper.backdoor:
+            poison_random(inputs, labels, 8, 1)
+
         optimizer.zero_grad()
 
         # forward + backward + optimize
@@ -53,11 +55,11 @@ def train(run_helper: ImageHelper, model: nn.Module, optimizer, criterion, write
         if i > 0 and i % run_helper.log_interval == 0:
             logger.info('[%d, %5d] loss: %.3f' %
                   (epoch + 1, i + 1, running_loss))
-            plot(writer, epoch * len(train_loader) + i, running_loss, 'Train Loss')
+            helper.plot(epoch * len(train_loader) + i, running_loss, 'Train Loss')
             running_loss = 0.0
 
 
-def test(run_helper: ImageHelper, model: nn.Module, criterion, writer, epoch):
+def test(run_helper: ImageHelper, model: nn.Module, criterion, epoch):
     model.eval()
     correct = 0
     total = 0
@@ -76,32 +78,31 @@ def test(run_helper: ImageHelper, model: nn.Module, criterion, writer, epoch):
             total += labels.size(0)
             correct += (predicted == labels).sum().item()
     main_acc = 100 * correct / total
-    logger.info(f'Epoch {epoch}. Accuracy: {main_acc}%')
-    plot(writer, x=epoch, y=main_acc, name="accuracy")
+    logger.warning(f'Epoch {epoch}. Accuracy: {main_acc}%')
+    helper.plot(x=epoch, y=main_acc, name="accuracy")
     return main_acc, total_loss
 
 
-def run(run_helper: ImageHelper, writer: SummaryWriter):
+def run(run_helper: ImageHelper):
 
     # load data
     run_helper.load_cifar10(helper.batch_size)
 
     # create model
-    model = models.resnet18(num_classes=len(run_helper.classes))
+    model = ResNet18(num_classes=len(run_helper.classes))
     model.to(run_helper.device)
+
     run_helper.check_resume_training(model)
 
     criterion = nn.CrossEntropyLoss().to(run_helper.device)
     optimizer = run_helper.get_optimizer(model)
     scheduler = torch.optim.lr_scheduler.MultiStepLR(optimizer, milestones=[150, 250, 350])
 
-
     for epoch in range(run_helper.start_epoch, helper.epochs+1):
-        train(run_helper, model, optimizer, criterion, writer=writer, epoch=epoch)
-        acc, loss = test(run_helper, model, criterion, writer=writer, epoch=epoch)
+        train(run_helper, model, optimizer, criterion, epoch=epoch)
+        acc, loss = test(run_helper, model, criterion, epoch=epoch)
         if run_helper.scheduler:
             scheduler.step(epoch)
-        writer.flush()
         run_helper.save_model(model, epoch, acc)
 
 
@@ -112,7 +113,6 @@ if __name__ == '__main__':
 
     args = parser.parse_args()
     d = datetime.now().strftime('%b.%d_%H.%M.%S')
-    wr = SummaryWriter(log_dir=f'runs/{args.name}')
 
     with open(args.params) as f:
         params = yaml.load(f)
@@ -124,32 +124,41 @@ if __name__ == '__main__':
         helper.corpus = torch.load(helper.params['corpus'])
         logger.info(helper.corpus.train.shape)
 
-    fh = logging.FileHandler(filename=f'{helper.folder_path}/log.txt')
-    formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    fh.setFormatter(formatter)
-    logger.addHandler(fh)
-    # logger.addHandler(logging.StreamHandler())
-    logger.setLevel(logging.DEBUG)
-    logger.info(f'current path: {helper.folder_path}')
+    if helper.log:
+        wr = SummaryWriter(log_dir=f'runs/{args.name}')
+        helper.writer = wr
+        logger = create_logger()
+        fh = logging.FileHandler(filename=f'{helper.folder_path}/log.txt')
+        formatter = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+        fh.setFormatter(formatter)
+        logger.addHandler(fh)
 
-    table = create_table(helper.params)
-    wr.add_text('Model Params', table)
+        logger.warning(f'Logging things. current path: {helper.folder_path}')
 
-    helper.params['tb_name'] = args.name
-    with open(f'{helper.folder_path}/params.yaml.txt', 'w') as f:
-        yaml.dump(helper.params, f)
+        table = create_table(helper.params)
+        helper.writer.add_text('Model Params', table)
+
+        helper.params['tb_name'] = args.name
+        with open(f'{helper.folder_path}/params.yaml.txt', 'w') as f:
+            yaml.dump(helper.params, f)
+    else:
+        logger = create_logger()
+
     try:
-        run(helper, wr)
-        print(f'You can find files in {helper.folder_path}. TB graph: {args.name}')
+        run(helper)
+        if helper.log:
+            print(f'You can find files in {helper.folder_path}. TB graph: {args.name}')
     except KeyboardInterrupt:
-        wr.flush()
-        answer = prompt('\nDelete the repo? (y/n): ')
-        if answer in ['Y', 'y', 'yes']:
 
-            shutil.rmtree(helper.folder_path)
-            shutil.rmtree(f'runs/{args.name}')
-            print(f"Fine. Deleted: {helper.folder_path}")
+        if helper.log:
+            answer = prompt('\nDelete the repo? (y/n): ')
+            if answer in ['Y', 'y', 'yes']:
+                logger.error(f"Fine. Deleted: {helper.folder_path}")
+                shutil.rmtree(helper.folder_path)
+                shutil.rmtree(f'runs/{args.name}')
+            else:
+                logger.error(f"Aborted training. Results: {helper.folder_path}. TB graph: {args.name}")
         else:
-            logger.info(f"Aborted training. Results: {helper.folder_path}. TB graph: {args.name}")
-    wr.flush()
+            logger.error(f"Aborted training. No output generated.")
+
